# packages/contracts/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/copropiedades/{id}/equipos": {
      "get": {
        "operationId": "EquiposController_listar",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EquiposDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Equipos de la copropiedad, sin credenciales",
        "tags": [
          "equipos"
        ]
      },
      "post": {
        "operationId": "EquiposController_crear",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AltaDeEquipoDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EquipoDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Da de alta un equipo; el secreto se guarda cifrado",
        "tags": [
          "equipos"
        ]
      }
    },
    "/copropiedades/{id}/equipos/prueba-de-conexion": {
      "post": {
        "operationId": "EquiposController_probar",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AltaDeEquipoDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultadoDeSondeoDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Prueba la conexión DESDE EL SERVIDOR, sin guardar nada",
        "tags": [
          "equipos"
        ]
      }
    },
    "/copropiedades/{id}/equipos/{equipoId}": {
      "put": {
        "operationId": "EquiposController_editar",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "equipoId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EdicionDeEquipoDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EquipoDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Edita un equipo. Sin «secreto» en el cuerpo, la clave no cambia",
        "tags": [
          "equipos"
        ]
      }
    },
    "/copropiedades/{id}/equipos/{equipoId}/baja": {
      "post": {
        "operationId": "EquiposController_desactivar",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "equipoId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BajaDeEquipoDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EquipoDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Baja lógica con motivo. Nunca borrado físico (RN-19)",
        "tags": [
          "equipos"
        ]
      }
    },
    "/copropiedades/{id}/equipos/{equipoId}/correcciones": {
      "post": {
        "operationId": "EquiposController_corregir",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "equipoId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CorreccionDeEquipoDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultadoDeCorreccionDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Corrige un campo del equipo. Exige confirmación y deja constancia",
        "tags": [
          "equipos"
        ]
      }
    },
    "/copropiedades/{id}/equipos/{equipoId}/diagnostico": {
      "post": {
        "operationId": "EquiposController_diagnosticar",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "equipoId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultadoDeSondeoDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Sondea un equipo en servicio con su clave guardada y devuelve su ficha",
        "tags": [
          "equipos"
        ]
      }
    },
    "/copropiedades/{id}/equipos/{equipoId}/reactivacion": {
      "post": {
        "operationId": "EquiposController_reactivar",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "equipoId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EquipoDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Vuelve a poner en servicio un equipo dado de baja",
        "tags": [
          "equipos"
        ]
      }
    },
...
  },
  "info": {
    "title": "Next Control Residencial — API",
    "description": "Next Control decide. El hardware ejecuta.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "AltaDeEquipoDto": {
        "type": "object",
        "properties": {
          "nombre": {
            "type": "string",
            "maxLength": 120
          },
          "tipo": {
            "type": "string",
            "enum": [
              "camara_lpr",
              "terminal_facial",
              "intercom",
              "rele",
              "controlador_io"
            ]
          },
          "host": {
            "type": "string",
            "example": "203.0.113.10"
          },
          "puerto": {
            "type": "number",
            "example": 80
          },
          "protocolo": {
            "type": "string",
            "enum": [
              "http",
              "https"
            ],
            "default": "http"
          },
          "usuario": {
            "type": "string",
            "maxLength": 64
          },
          "secreto": {
            "type": "string",
            "maxLength": 128,
            "writeOnly": true
          },
          "canalBarrera": {
            "type": "number"
          },
          "numeroDePuerta": {
            "type": "number"
          },
          "canalDeAudio": {
            "type": "number"
          },
          "fabricante": {
            "type": "string",
            "maxLength": 80
          },
          "modoDeTerminal": {
            "type": "string",
            "enum": [
              "reporta_y_espera",
              "decide_el_equipo"
            ]
          },
          "canalDeAudioHabilitado": {
            "type": "boolean",
            "default": false
          },
          "probarConexion": {
            "type": "boolean",
            "default": true
          }
        },
        "required": [
          "nombre",
          "tipo",
          "host",
          "puerto",
          "protocolo",
          "usuario"
        ]
      },
...
      "BajaDeEquipoDto": {
        "type": "object",
        "properties": {
          "motivo": {
            "type": "string",
            "maxLength": 300
          }
        },
        "required": [
          "motivo"
        ]
      },
...
      "EdicionDeEquipoDto": {
        "type": "object",
        "properties": {
          "nombre": {
            "type": "string",
            "maxLength": 120
          },
          "tipo": {
            "type": "string",
            "enum": [
              "camara_lpr",
              "terminal_facial",
              "intercom",
              "rele",
              "controlador_io"
            ]
          },
          "host": {
            "type": "string",
            "example": "203.0.113.10"
          },
          "puerto": {
            "type": "number",
            "example": 80
          },
          "protocolo": {
            "type": "string",
            "enum": [
              "http",
              "https"
            ],
            "default": "http"
          },
          "usuario": {
            "type": "string",
            "maxLength": 64
          },
          "secreto": {
            "type": "string",
            "maxLength": 128,
            "writeOnly": true
          },
          "canalBarrera": {
            "type": "number"
          },
          "numeroDePuerta": {
            "type": "number"
          },
          "canalDeAudio": {
            "type": "number"
          },
          "fabricante": {
            "type": "string",
            "maxLength": 80
          },
          "modoDeTerminal": {
            "type": "string",
            "enum": [
              "reporta_y_espera",
              "decide_el_equipo"
            ]
          },
          "canalDeAudioHabilitado": {
            "type": "boolean",
            "default": false
          },
          "probarConexion": {
            "type": "boolean",
            "default": true
          }
        }
      },
...
      "EquipoDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "nombre": {
            "type": "string"
          },
          "tipo": {
            "type": "string",
            "enum": [
              "camara_lpr",
              "terminal_facial",
              "intercom",
              "rele",
              "controlador_io"
            ]
          },
          "modelo": {
            "type": "string",
            "nullable": true
          },
          "firmware": {
            "type": "string",
            "nullable": true
          },
          "canalBarrera": {
            "type": "number",
            "nullable": true
          },
          "numeroDePuerta": {
            "type": "number",
            "nullable": true
          },
          "canalDeAudio": {
            "type": "number",
            "nullable": true
          },
          "fabricante": {
            "type": "string",
            "nullable": true
          },
          "modoDeTerminal": {
            "type": "string",
            "nullable": true,
            "enum": [
              "reporta_y_espera",
              "decide_el_equipo"
            ]
          },
          "canalDeAudioHabilitado": {
            "type": "boolean"
          },
          "capacidades": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/CapacidadesDeEquipoDto"
              }
            ]
          },
          "verificacion": {
            "type": "string",
            "enum": [
              "no_verificado",
              "verificado",
              "rechazado"
            ]
          },
          "verificadoEn": {
            "type": "string",
            "nullable": true
          },
          "motivoNoVerificado": {
            "type": "string",
            "nullable": true
          },
          "estado": {
            "type": "string",
            "enum": [
              "activo",
              "inactivo"
            ]
          }
        },
        "required": [
          "id",
          "nombre",
          "tipo",
          "modelo",
          "firmware",
          "canalBarrera",
          "numeroDePuerta",
          "canalDeAudio",
          "fabricante",
          "modoDeTerminal",
          "canalDeAudioHabilitado",
          "capacidades",
          "verificacion",
          "verificadoEn",
          "motivoNoVerificado",
          "estado"
        ]
      },
      "EquiposDto": {
        "type": "object",
        "properties": {
          "equipos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EquipoDto"
            }
          }
        },
        "required": [
          "equipos"
        ]
      },
...
      "ResultadoDeSondeoDto": {
        "type": "object",
        "properties": {
          "clase": {
            "type": "string",
            "enum": [
              "alcanzado",
              "decide_solo",
              "credencial",
              "inalcanzable"
            ],
            "description": "Cuatro resultados distintos, nunca uno genérico: cada uno se resuelve de una manera."
          },
          "detalle": {
            "type": "string"
          },
          "modelo": {
            "type": "string",
            "nullable": true
          },
          "firmware": {
            "type": "string",
            "nullable": true
          },
          "latenciaMs": {
            "type": "number",
            "nullable": true
          },
          "verificado": {
            "type": "boolean"
          },
          "ficha": {
            "description": "Qué hay que cambiar en el equipo, campo por campo. Ausente cuando no se sondeó: la falta de ficha no es una ficha vacía.",
            "allOf": [
              {
                "$ref": "#/components/schemas/FichaDelEquipoDto"
              }
            ]
          },
          "capacidades": {
            "description": "Lo que el equipo declaró poder hacer. Ausente cuando no se alcanzó.",
            "allOf": [
              {
                "$ref": "#/components/schemas/CapacidadesDeEquipoDto"
              }
            ]
          }
        },
        "required": [
          "clase",
          "detalle",
          "modelo",
          "firmware",
          "latenciaMs",
          "verificado"
        ]
      },
...
    }
  }
}